# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fnashha Home Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── GOVERNORATES ────────────────────────────────────────────────────────────
...
  /ratings:
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
    post:
      operationId: createRating
      tags: [ratings]
      summary: Rate a technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /ratings/{id}:
    delete:
      operationId: deleteRating
      tags: [ratings]
      summary: Delete rating (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── POINTS ──────────────────────────────────────────────────────────────────
...
  /support/tickets:
    get:
      operationId: listTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /support/tickets/{id}:
    get:
      operationId: getTicket
      tags: [support]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [support]
      summary: Update ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /support/tickets/{id}/reply:
    post:
      operationId: replyTicket
      tags: [support]
      summary: Reply to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketReply"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, mobile, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, technician, admin, super_admin]
        status:
          type: string
          enum: [active, pending, suspended, banned, rejected]
        profileImage:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
        technicianProfile:
          $ref: "#/components/schemas/TechnicianProfile"
...
    TechnicianProfile:
      type: object
      required: [id, userId, nationalId, approvalStatus]
      properties:
        id:
          type: integer
        userId:
          type: integer
        nationalId:
          type: string
        personalPhoto:
          type: ["string", "null"]
        nationalIdFront:
          type: ["string", "null"]
        nationalIdBack:
          type: ["string", "null"]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        pointsBalance:
          type: integer
        averageRating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        areas:
          type: array
          items:
            $ref: "#/components/schemas/Area"
        primaryArea:
          $ref: "#/components/schemas/Area"
...
    Service:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
    ServiceInput:
      type: object
      required: [name, nameAr]
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer

    # ─── GOVERNORATE ───────────────────────────────────────────────────────────
    Governorate:
      type: object
      required: [id, name, nameAr, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
...
    Area:
      type: object
      required: [id, name, nameAr, governorateId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        governorate:
          $ref: "#/components/schemas/Governorate"
        isActive:
          type: boolean
...
    Rating:
      type: object
      required: [id, requestId, customerId, technicianId, stars, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        customerId:
          type: integer
        customer:
          $ref: "#/components/schemas/User"
        technicianId:
          type: integer
        technician:
          $ref: "#/components/schemas/User"
        stars:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: ["string", "null"]
        createdAt:
          type: string
    RatingInput:
      type: object
      required: [requestId, technicianId, stars]
      properties:
        requestId:
          type: integer
        technicianId:
          type: integer
        stars:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string

    # ─── POINTS ────────────────────────────────────────────────────────────────
...
    Ticket:
      type: object
      required: [id, userId, subject, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/TicketReply"
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
        message:
          type: string
        images:
          type: array
          items:
            type: string
    TicketUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
    TicketReply:
      type: object
      required: [id, ticketId, senderId, message, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        message:
          type: string
        createdAt:
          type: string
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── NOTIFICATION ──────────────────────────────────────────────────────────
...